dialog-0b01_0_12
0b/01 Lib 0.12 (Manual revision 1):

This is the first public release of Dialog.

Dialog is currently in its beta stage, which means that the language
may still undergo significant changes. It also means that you, dear
potential story author, still have a substantial chance to influence
what those changes will be.

The source code for the compiler is currently rather messy, and I'm
planning a major clean-up. However, it should be portable, and it works
according to the language specification (as far as I know).