dialog-0d01_0_19
0d/01 Lib 0.19 (Manual revision 4):

Introduced the Interactive Debugger, with corresponding
modifications to the documentation.

Added '(breakpoint)' built-in predicate.

Library: Added '(actions on)', '(actions off)', and '(scope)'
predicates to be queried from the debugger. The corresponding
player-accessible commands remain in the debugging extension.

Library: Modified the treatment of UNDO and AGAIN, to better
support the interactive debugger.

Library: Minor improvements.