dialog-0g06_0_29
0g/06 Lib 0.29 (Manual revision 14):

Library: In the before-rules for eating and drinking, only
attempt to pick up the indicated object if it is edible or
potable.

Library: Treat "X, tell me about Y" as "ask X about Y".

Library: Fixed a bug where '(describe topic $)' couldn't deal
with ambiguity.

Compiler: Fixed a bug related to the optimization of nested
disjunctions.