dialog-0j04_0_38
0j/04, Lib 0.38 (Manual revision 23):

Library: Undo is performed as an action, [undo], and reported
via '(narrate undoing)'. It is still parsed as a special case.

Library: When the player types an object name in response to a
direct question (e.g. "To whom?"), don't understand it as a
request to perform the default action (e.g. examine).

Compiler: Fixed a bug that disallowed @-prefixed words in
slash-expressions.

Compiler: Several improvements to the optimizer.